adding the python code
This is the main python code that I wrote in vscode to finish the module challenge. 
I've missed one of the challenge as I wasn't able to figure out the formula for it in the part 1. But that's the only one that I haven't been able to solve in time.